refactor-#29: add index.ts to dtos
