+++ src/customer/dto/index.ts
@@ -1,0 +1,4 @@
+export * from './create-customer-profile.dto';
+export * from './get-customer-profile.dto';
+export * from './remove-customer-profile.dto';
+export * from './update-customer-profile.dto';
