+++ src/customer/services/customer.service.ts
@@ -6,12 +6,9 @@ import {
   RemoveCustomerProfileCommand,
   UpdateCustomerProfileCommand,
 } from '../commands/impl';
-import { CreateCustomerProfileDto } from '../dto/create-customer-profile.dto';
-import { GetCustomerProfileDto } from '../dto/get-customer-profile.dto';
+import { CreateCustomerProfileDto, RemoveCustomerProfileDto, GetCustomerProfileDto, UpdateCustomerProfileDto } from '../dto';
 import { CustomerEntity } from '../entities';
 import { GetCustomerProfileQuery } from '../queries/impl';
-import { RemoveCustomerProfileDto } from '../dto/remove-customer-profile.dto';
-import { UpdateCustomerProfileDto } from '../dto/update-customer-profile.dto';
 
 @Injectable()
 export class CustomerService {
